+++ POS.Data/Entities/Role.cs
@@ -14,8 +14,8 @@ namespace POS.Data.Entities
         public int? CompanyId { get; set; }
         public DateTime? CreateDate { get; set; }
         public DateTime? LastModifyDate { get; set; }
-        public int? InsertedBy { get; set; }
-        public int? ModifiedBy { get; set; }
+        public long? InsertedBy { get; set; }
+        public long? ModifiedBy { get; set; }
 
     }
 }
